Initial commit
So I was doing work at the SubmittingFiles branch. I updated it from  master to prepare a pull req. Alex had added a common TestProjectFile and I was gonna copy over my tests from my Tests folder and scrap it afterwards. It got really messed up. Conflicts in the autogenerated files that i could not understand or solve. Tried multiple workarounds but to no use. So I abandoned the SubmittingFiles branch, created a new branch called SubmittingFilesNew and just copied over the actual code i wanted to keep to that one instead. Still I can see in the folders view that there still is this folder Tests which I don't understand where it comes from because I made a brand new branch from master which should have nothing to do with the corrupted old branch SubmittingFiles. But everything works so I decided to let it be now.