Refactoring the SearchEngine to make it more readable
Extracting to smaller methods so that one method truly is responsible for one thing
Also tried to make SearchEngine vs FilePathVerifier more abstract and portable. FilePathVerifier is no longer responsible for deciding what formats it should evaluate, it is given this as a parameter, making it more abstract and usable in other contexts. The SearchEngine is now responsible for keeping track if file formats and this is decided in the constructor, meaning you could theoretically use our SearchEngine to process other file formats